# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "FriendshipRequest": {
        "properties": {
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "user1_id": {
            "default": 0,
            "title": "User1 Id",
            "type": "integer"
          },
          "user2_id": {
            "default": 0,
            "title": "User2 Id",
            "type": "integer"
          }
        },
        "title": "FriendshipRequest",
        "type": "object"
      },
      "FriendshipResponse": {
        "properties": {
          "action_user_id": {
            "title": "Action User Id",
            "type": "integer"
          },
          "status": {
            "default": "",
            "title": "Status",
            "type": "string"
          },
          "user1_id": {
            "default": 0,
            "title": "User1 Id",
            "type": "integer"
          },
          "user2_id": {
            "default": 0,
            "title": "User2 Id",
            "type": "integer"
          }
        },
        "required": [
          "user1_id",
          "user2_id",
          "status",
          "action_user_id"
        ],
        "title": "FriendshipResponse",
        "type": "object"
      },
...
      "LeaderboardResponse": {
        "properties": {
          "max_weight": {
            "title": "Max Weight",
            "type": "number"
          },
          "percentile": {
            "title": "Percentile",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "max_weight",
          "rank",
          "percentile"
        ],
        "title": "LeaderboardResponse",
        "type": "object"
      },
...
      "MuscleRankResponse": {
        "properties": {
          "exercise_key": {
            "title": "Exercise Key",
            "type": "string"
          },
          "exercise_name": {
            "title": "Exercise Name",
            "type": "string"
          },
          "max_weight": {
            "title": "Max Weight",
            "type": "number"
          },
          "percentile": {
            "title": "Percentile",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "source_workout_item_id": {
            "title": "Source Workout Item Id",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "exercise_key",
          "exercise_name",
          "max_weight",
          "source_workout_item_id",
          "rank",
          "percentile"
        ],
        "title": "MuscleRankResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "MetricsBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "GopherFit API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/social/friendships": {
      "get": {
        "operationId": "friendships_social_friendships_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Friendships Social Friendships Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Friendships",
        "tags": [
          "social"
        ]
      },
      "post": {
        "operationId": "create_friendship_social_friendships_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendshipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "User not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Friendship already exists"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Friendship",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/accepted": {
      "get": {
        "operationId": "accepted_social_friendships_accepted_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Accepted Social Friendships Accepted Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accepted",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/inblocks": {
      "get": {
        "operationId": "incoming_blocks_social_friendships_inblocks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Incoming Blocks Social Friendships Inblocks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incoming Blocks",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/inpending": {
      "get": {
        "operationId": "incoming_requests_social_friendships_inpending_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Incoming Requests Social Friendships Inpending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incoming Requests",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/outblocks": {
      "get": {
        "operationId": "outgoing_blocks_social_friendships_outblocks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Outgoing Blocks Social Friendships Outblocks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Outgoing Blocks",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/outpending": {
      "get": {
        "operationId": "outgoing_requests_social_friendships_outpending_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  },
                  "title": "Response Outgoing Requests Social Friendships Outpending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Outgoing Requests",
        "tags": [
          "social"
        ]
      }
    },
    "/social/friendships/{user2_id}": {
      "delete": {
        "operationId": "delete_friendship_social_friendships__user2_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user2_id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "User2 Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Relationship changed; reload before retrying"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Friendship",
        "tags": [
          "social"
        ]
      },
      "get": {
        "operationId": "friendship_social_friendships__user2_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user2_id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "User2 Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Friendship",
        "tags": [
          "social"
        ]
      },
      "put": {
        "operationId": "update_friendship_social_friendships__user2_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "user2_id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "User2 Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendshipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Relationship changed; reload before retrying"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Friendship",
        "tags": [
          "social"
        ]
      }
    },
    "/social/leaderboard": {
      "get": {
        "operationId": "leaderboard_social_leaderboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "exercise",
            "required": false,
            "schema": {
              "default": "",
              "title": "Exercise",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LeaderboardResponse"
                  },
                  "title": "Response Leaderboard Social Leaderboard Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Leaderboard",
        "tags": [
          "social"
        ]
      }
    },
    "/social/muscle-ranks": {
      "get": {
        "operationId": "muscle_ranks_social_muscle_ranks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MuscleRankResponse"
                  },
                  "title": "Response Muscle Ranks Social Muscle Ranks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Muscle Ranks",
        "tags": [
          "social"
        ]
      }
    },
...
  }
}